client - fix bug and some refactoring

+++ client/utils.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include "utils.hpp"
 
-void printBytes(const uint8_t data[], int data_size) {
+void printBytes(const uint8_t data[], size_t data_size) {
     for (int i =0; i < data_size; i++) {
         unsigned char byte = data[i];
         std::cout << std::hex << (int)byte << " ";
