+++ client/utils.hpp
@@ -1,4 +1,4 @@
 #pragma once
 #include <cstdint>
 
-void printBytes(const uint8_t data[], int data_size);
+void printBytes(const uint8_t data[], size_t data_size);
